Back/requirements.txt
rdflib
python-chess
pandas
transformers
datasets
sklearn
huggingface_hub
jsonlines
peft
wandb 
onnx
wandb
